Nearly fully automate the process
Now I do not need to clear out form data until google complains that I
have too many files. Unfortunately, google api does not easily allow
deletion but may look into it if it becomes a problem.

Now all I need is to get this on a server with Cron and I will never
have to think about it again.