Complete major refactor
Split the monolith file into separate files that specialise in
interfacing with forms (at the moment still Google Forms) and the SMTP
email client.

During this I have also introduced type hints to verify that the
accessing config items is sound.

This is the penultimate step to getting a cronjob enabled script. The
final step is to remove any manual entry points mainly in the custom
header script. It will be sad to lose them but that is the price of
efficint automation.