Initial server commit
This will be hosted on [my website](https://atp45.user.srcf.net). For
now this is all just pretty things with no brain, the himbo of the form
server if you will.

This branch will eventually become the main branch an a separate `google` branch will be saved to maintain usability for people who do not want to roll their own.

I have plans to make this into a fairly fledged product that people can
almost plug and play with.