Restructuring and publishing
This has been adjusted to interact with [my website](https://github.com/skyepurchase/website)
with minor adjustments made to the styling and references to make sure it works.

For now it is barely functional. The addition of a SQL database and full
integration with the existing code needs to occur.