Use the new database for authentication
The passcode is now stored in the database. By iterating over newsletter
entries the typed passcode can be compared against possible newsletters.
This is suboptimal as it requires newsletter passcodes to be unique. A
better system will potentially be implemented requiring the newsletter
name to be entered as well.

I create the passcodes so I can prevent duplicates for now.