Store all questions in database
Rather than just the user generated questions all the questions need to
be stored in the database. The 'text' or 'image' tag is added so that
the "Photo Wall" question can be rendered properly. The standard
questions are given the "base" tag. These now need to be added every
issue but also opens the door to more dynamic forms.

This does pave the way for more questions including image submissions
and the answer database can store an image for all possible answers so
definitely something on the horizon.