Add answer insertion
This avoids dealing with images just yet as those will require some file
management and decoding the post request. No problems seen so far,
though deduplicating the database would be best.

Additionally `utils/email.py` updated to use the `typing` library to
prevent crashes on the server, which is `python v3.8`, in the future.