Remove google forms reliance
The mailer no longer interacts with Google forms! Every email is simply
the same link to wherever the site is hosted.

There is some clean up that can be done as now the mailer is near
identical regardless of request type. The assumed config file has also
changed, eventually (once this and the google version are finalised), I
will create two READMEs that detail everything.