Utilise JWT authentication token
Rather than authenticating the user twice (when unlocking the newsletter
and when completing an action) the authentication is done at the higher
level and the CGI functions take in the post-authentication newsletter
data.

This way authentication can be handled by a single call to
`authenticate` and a JSON web token can be generated. This can then be
repeatedly verified and store the relevant newsletter data (without
storing the passcode!!).